Migrate the MentoringForm tests. This component is currently unused, but may be rendered again in the future.
